refactor: registration into specific package
# Conflicts:
#	search-service/src/main/kotlin/com/egm/stellio/search/service/registration/service/ServiceRegistrationService.kt
