+++ search-service/src/test/kotlin/com/egm/stellio/search/service/model/ServiceRegistrationTests.kt
@@ -1,5 +1,8 @@
 package com.egm.stellio.search.service.model
 
+import com.egm.stellio.search.service.registration.model.InputInformationType
+import com.egm.stellio.search.service.registration.model.ServiceMode
+import com.egm.stellio.search.service.registration.model.ServiceRegistration
 import com.egm.stellio.shared.model.BadRequestDataException
 import com.egm.stellio.shared.util.JsonUtils.deserializeAsMap
 import com.egm.stellio.shared.util.shouldFailWith
